- Restructuring project architecture to make better use of the MVVM architecture - Added new viewmodels and arranged files into appropriate packages - Began code cleanup and refactoring
