+++ .idea/misc.xml
@@ -32,10 +32,13 @@
         <entry key="..\:/Users/eire1/AndroidStudioProjects/MultiAlarmClock/app/src/main/res/layout/iv_refresh.xml" value="0.11499493414387031" />
         <entry key="..\:/Users/eire1/AndroidStudioProjects/MultiAlarmClock/app/src/main/res/layout/toolbar.xml" value="0.17255434782608695" />
         <entry key="..\:/Users/eire1/AndroidStudioProjects/MultiAlarmClock/app/src/main/res/menu/menu_main.xml" value="0.23645833333333333" />
+        <entry key="app/src/main/res/layout/activity_build_interval_alarm.xml" value="0.1" />
+        <entry key="app/src/main/res/layout/activity_main.xml" value="0.1296875" />
+        <entry key="app/src/main/res/menu/menu_main.xml" value="0.12777777777777777" />
       </map>
     </option>
   </component>
-  <component name="ProjectRootManager" version="2" languageLevel="JDK_11" default="true" project-jdk-name="1.8" project-jdk-type="JavaSDK">
+  <component name="ProjectRootManager" version="2" languageLevel="JDK_1_8" default="true" project-jdk-name="1.8" project-jdk-type="JavaSDK">
     <output url="file://$PROJECT_DIR$/build/classes" />
   </component>
   <component name="ProjectType">
