+++ app/src/main/AndroidManifest.xml
@@ -1,18 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     package="com.example.multialarmclock">
 
     <application
+        android:name=".App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.AppCompat.Light.NoActionBar"
-        android:name="com.example.multialarmclock.App">
+        android:theme="@style/Theme.AppCompat.Light.NoActionBar">
         <activity
-            android:name=".MainActivity"
+            android:name=".feature.alarmIntervalBuilder.BuildIntervalAlarmActivity"
+            android:exported="false" />
+        <activity
+            android:name=".feature.homeScreen.HomeActivity"
             android:exported="true"
             android:label="@string/title_activity_build_new_alarm"
             android:theme="@style/AppTheme">
@@ -22,13 +24,11 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".BuildNewAlarm" />
+        <activity android:name=".feature.alarmIntervalBuilder.BuildNewAlarm" />
 
-        <receiver android:name=".AlarmReceiver"
-        android:exported="false"/>
-    
+        <receiver
+            android:name=".AlarmReceiver"
+            android:exported="false" />
     </application>
-    
-    
 
 </manifest>
