--- app/src/main/java/com/example/multialarmclock/data/MultiAlarm.kt
@@ -1,4 +1,0 @@
-package com.example.multialarmclock.data
-
-class MultiAlarm {
-}
