+++ app/src/main/java/com/example/multialarmclock/feature/alarmIntervalBuilder/AlarmIntervalBuilder.kt
@@ -1,4 +1,4 @@
-package com.example.multialarmclock
+package com.example.multialarmclock.feature.alarmIntervalBuilder
 
 import android.app.Activity
 import android.content.Intent
@@ -18,6 +18,7 @@ import androidx.activity.result.contract.ActivityResultContracts
 import androidx.annotation.RequiresApi
 import androidx.cardview.widget.CardView
 import androidx.lifecycle.ViewModelProvider
+import com.example.multialarmclock.R
 import com.example.multialarmclock.classes.AlarmViewModel
 import com.example.multialarmclock.classes.TimeViewModel
 import com.example.multialarmclock.data.BuildNewAlarmModel
@@ -27,21 +28,13 @@ import kotlinx.coroutines.InternalCoroutinesApi
 import java.util.*
 import kotlin.collections.ArrayList
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
-
-
 /**
  * A simple [Fragment] subclass.
  * Use the [AlarmIntervalBuilder.newInstance] factory method to
  * create an instance of this fragment.
  */
 class AlarmIntervalBuilder : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
+
     private lateinit var binding: ActivityBuildNewAlarmBinding
     @RequiresApi(Build.VERSION_CODES.P)
     internal lateinit var slider: FluidSlider
@@ -82,13 +75,10 @@ class AlarmIntervalBuilder : Fragment() {
     var chosenRTUri: Uri?=null
     internal lateinit var intervalPicker: NumberPicker
 
-    @InternalCoroutinesApi
+
+    @OptIn(InternalCoroutinesApi::class)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
-        }
         mAlarmViewModel = ViewModelProvider(this).get(AlarmViewModel::class.java)
 
 
@@ -103,15 +93,10 @@ class AlarmIntervalBuilder : Fragment() {
             toggleOff.isChecked = false
         }
 
-
-
-
-
-
-
     }
 
-    @InternalCoroutinesApi
+
+    @OptIn(InternalCoroutinesApi::class)
     @RequiresApi(Build.VERSION_CODES.M)
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -218,23 +203,10 @@ class AlarmIntervalBuilder : Fragment() {
     }
 
     companion object {
-        /**
-         * Use this factory method to create a new instance of
-         * this fragment using the provided parameters.
-         *
-         * @param param1 Parameter 1.
-         * @param param2 Parameter 2.
-         * @return A new instance of fragment AlarmIntervalBuilder.
-         */
-        // TODO: Rename and change types and number of parameters
+
         @JvmStatic
-        fun newInstance(param1: String, param2: String) =
-            AlarmIntervalBuilder().apply {
-                arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
-                }
-            }
+        fun newInstance() = AlarmIntervalBuilder()
+
     }
 
     @RequiresApi(Build.VERSION_CODES.M)
