+++ app/src/main/java/com/example/multialarmclock/feature/homeScreen/HomeScreenViewModel.kt
@@ -1,0 +1,10 @@
+package com.example.multialarmclock.feature.homeScreen
+
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+
+@HiltViewModel
+class HomeScreenViewModel: ViewModel() {
+
+
+}
