+++ app/src/main/res/layout/activity_build_interval_alarm.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".feature.alarmIntervalBuilder.BuildIntervalAlarmActivity">
+
+    <androidx.fragment.app.FragmentContainerView
+        android:id="@+id/build_interval_alarm_frag"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:name="com.example.multialarmclock.feature.alarmIntervalBuilder.AlarmIntervalBuilder">
+
+    </androidx.fragment.app.FragmentContainerView>
+
+</androidx.constraintlayout.widget.ConstraintLayout>
