+++ app/src/main/res/menu/menu_build_new_alarm.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.example.multialarmclock.BuildNewAlarm">
+    tools:context="com.example.multialarmclock.feature.alarmIntervalBuilder.BuildNewAlarm">
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
