+++ app/src/main/res/menu/menu_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".feature.homeScreen.HomeActivity">
 
     <item
         android:id="@+id/settingsid"
