+++ build.gradle
@@ -8,7 +8,9 @@ buildscript {
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:7.0.1'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.30"
+        classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.10'
+        classpath 'com.google.dagger:hilt-android-gradle-plugin:2.38.1'
+
 //        classpath 'com.squareup.sqldelight:gradle-plugin:1.5.2'
 
         // NOTE: Do not place your application dependencies here; they belong
